+++ src/editing/tools/Tools.js
@@ -15,6 +15,7 @@ import Transform from './Transform.js';
 import ClayBuildup from './ClayBuildup.js';
 import ZSphereTool from './ZSphereTool.js';
 import Topology from './Topology.js';
+import MeasureTool from '../../measure/MeasureTool.js';
 
 var Tools = [];
 
@@ -34,6 +35,7 @@ Tools[Enums.Tools.TRANSFORM] = Transform;
 Tools[Enums.Tools.CLAYBUILDUP] = ClayBuildup;
 Tools[Enums.Tools.ZSPHERE] = ZSphereTool;
 Tools[Enums.Tools.TOPOLOGY] = Topology;
+Tools[Enums.Tools.MEASURE] = MeasureTool;
 
 Tools[Enums.Tools.BRUSH].uiName = 'sculptBrush';
 Tools[Enums.Tools.INFLATE].uiName = 'sculptInflate';
@@ -51,5 +53,6 @@ Tools[Enums.Tools.TRANSFORM].uiName = 'sculptTransform';
 Tools[Enums.Tools.CLAYBUILDUP].uiName = 'sculptClayBuildup';
 Tools[Enums.Tools.ZSPHERE].uiName = 'sculptZSphere';
 Tools[Enums.Tools.TOPOLOGY].uiName = 'sculptTopology';
+Tools[Enums.Tools.MEASURE].uiName = 'measureTitle';
 
 export default Tools;